backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.1
pillow==11.1.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
scipy
colour-science==0.4.6
anthropic==0.43.0
openai==1.58.1
python-multipart==0.0.20
aiofiles==24.1.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
